refactor(dts): separate public and internal types
